Use express-generator to organize project
