+++ package.json
@@ -6,10 +6,14 @@
     "start": "node ./bin/www"
   },
   "dependencies": {
-    "express": "^4.17.1",
+    "cookie-parser": "~1.4.4",
+    "debug": "~2.6.9",
+    "express": "~4.16.1",
     "jquery": "^3.5.1",
     "pg": "^8.4.1",
     "pug": "^3.0.0",
-    "puppeteer": "^5.3.0"
+    "puppeteer": "^5.3.0",
+    "http-errors": "~1.6.3",
+    "morgan": "~1.9.1"
   }
 }
