+++ public/stylesheets/style.css
@@ -1,3 +1,12 @@
+body {
+  padding: 50px;
+  font: 14px "Lucida Grande", Helvetica, Arial, sans-serif;
+}
+
+a {
+  color: #00B7FF;
+}
+
 .invisble {
-    display:none
+  display:none
 }
