+++ routes/index.js
@@ -3,7 +3,7 @@ var router = express.Router();
 
 /* GET home page. */
 router.get('/', function(req, res, next) {
-  res.render('index', { title: 'Express' });
+  res.render('configSearch', { title: 'Moc Meetup' });
 });
 
 module.exports = router;
